# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: School Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teachers:
    get:
      operationId: getTeachers
      tags: [teachers]
      summary: List all teachers
      responses:
        "200":
          description: List of teachers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teacher"
    post:
      operationId: createTeacher
      tags: [teachers]
      summary: Create a teacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeacher"
      responses:
        "201":
          description: Created teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
  /teachers/{id}:
    put:
      operationId: updateTeacher
      tags: [teachers]
      summary: Update a teacher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeacher"
      responses:
        "200":
          description: Updated teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
    delete:
      operationId: deleteTeacher
      tags: [teachers]
      summary: Delete a teacher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /timetable:
    get:
      operationId: getTimetable
      tags: [timetable]
      summary: Get full timetable
      responses:
        "200":
          description: List of timetable slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimetableSlot"
    post:
      operationId: createTimetableSlot
      tags: [timetable]
      summary: Create a timetable slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimetableSlot"
      responses:
        "201":
          description: Created slot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimetableSlot"
  /timetable/{id}:
    put:
      operationId: updateTimetableSlot
      tags: [timetable]
      summary: Update a timetable slot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimetableSlot"
      responses:
        "200":
          description: Updated slot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimetableSlot"
    delete:
      operationId: deleteTimetableSlot
      tags: [timetable]
      summary: Delete a timetable slot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Teacher:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        subject:
          type: string
        maxWeeklyHours:
          type: integer
        weeklySubstitutionCount:
          type: integer
      required: [id, name, subject, maxWeeklyHours, weeklySubstitutionCount]
    CreateTeacher:
      type: object
      properties:
        name:
          type: string
        subject:
          type: string
        maxWeeklyHours:
          type: integer
      required: [name, subject, maxWeeklyHours]
...
    TimetableSlot:
      type: object
      properties:
        id:
          type: integer
        dayOfWeek:
          type: integer
          description: "0=Sunday, 1=Monday, ..., 4=Thursday"
        period:
          type: integer
          description: "1-8"
        teacherId:
          type: integer
        classId:
          type: integer
        subjectId:
          type: integer
        teacherName:
          type: string
        className:
          type: string
        subjectName:
          type: string
      required: [id, dayOfWeek, period, teacherId, classId, subjectId, teacherName, className, subjectName]
    CreateTimetableSlot:
      type: object
      properties:
        dayOfWeek:
          type: integer
        period:
          type: integer
        teacherId:
          type: integer
        classId:
          type: integer
        subjectId:
          type: integer
      required: [dayOfWeek, period, teacherId, classId, subjectId]
...