# lib/api-zod/src/generated/types/timetableSlot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * School Management API
 * OpenAPI spec version: 0.1.0
 */

export interface TimetableSlot {
  id: number;
  /** 0=Sunday, 1=Monday, ..., 4=Thursday */
  dayOfWeek: number;
  /** 1-8 */
  period: number;
  teacherId: number;
  classId: number;
  subjectId: number;
  teacherName: string;
  className: string;
  subjectName: string;
}